add support for searching for alt titles
if a match is not found based on the main title, this will now do
follow-up searches for the alt titles as well.

This also fixes a bug where the index was global and wasn't respecting
deletes / renames of files on disk.  Now it creates a new index per
run and passes it around.