change preprocess_name logic to eliminate all spaces
previously it was normalizing to space-separated words.
now this will concat everything together.  the impact is that for
items that after normalization that had < 3 characters this will
now concat the next word - and still allow for a 3 char prefix.
this helps to fix some edge cases and at worst will expand the match
pool slightly for a small subset of media