remove extraction of release years.
This was causing bad matches due to physical releases sometimes lagging
by years and then hitting other items with similar names.  This is likely
no longer needed due to grabbing the secondaryYear from arr now.

This also grabs the year from arr and puts it onto the object in
case we want to leverage it later.