if there's no edition, set a default value for upload purposes
this allows something that has an edition to be upgraded (quality
score-wise) to a version that doesn't have an edition.  Previously
if an edition existed first but then an upgrade did not have an
edition the code would just look at the libraries it was added to.
now this will also see if an edition is there for the default case.