Remove upload data for orphaned assets that were previously uploaded
Some people remove shows as soon as they are watched and this can
lead to situations where we have uploaded a poster previously and
the show still exists in Sonarr so we will not remove the cache data.
This change checks for that case and for items that were previously
uploaded and now the season exists but has no files, it will remove
the upload data.